### 2019-04-05 #292 @noamross
citesdb: A high-performance database of shipment-level CITES trade data
Submitting Author: Noam Ross (@noamross)
Repository: https://github.com/ecohealthalliance/citesdb
Version submitted:   0.1.0
Archive: https://zenodo.org/record/3234871
Version accepted: 0.2.0

---

```
Package: citesdb
Version: 0.1.0
Title: A high-performance database of shipment-level CITES trade data
Description: Provides convenient access to over 40 years and 20 million records of
    endangered wildlife trade data from the Convention on International Trade
    in Endangered Species of Wild Fauna and Flora, stored on a local on-disk,
    out-of memory 'MonetDB' database for bulk analysis.
    person("Noam", "Ross", , "ross@ecohealthalliance.org", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-2136-0000")),
    person("Evan A.", "Eskew", , "eskew@ecohealthalliance.org", role = c("aut"), comment = c(ORCID = "0000-0002-1153-5356")),
    person("EcoHealth Alliance", role = "cph"))
  MonetDBLite,
  rappdirs,
  DBI,
  httr,
  R.utils,
  tools,
  dbplyr
  roxygen2,
  rstudioapi,
  lintr,
  callr,
  here,
  rcites
```

**citesdb** gives researchers access to the large CITES wildlife trade dataset by syncing it to a local out-of-memory trade database and provides an interface to that local database.

Researchers and conservationists studying the wildlife trade.

Our deprecated [**cites**](https://github.com/ecohealthalliance/cites) package used a different framework to access a smaller subset of the data.  The [**rcites**](https://github.com/ropensci/rcites) package provides complementary data.

_I note the relatively low test coverage (~65%) is due to a verbose section of code configuring the RStudio interactive interface.  Other than this, the remaining code has >90% coverage._

    - [x] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.2630836

_Recommended reviewers_:  **karinorman** will be familiar with the database used here via her work on **taxadb**, **JonasGeschke** and **KevCaz** will know about the data and their application through their work on **rcites**.  Feedback from someone involved in metadata projects - e.g.: Bryce Mecum, Auriel Fournier, Kelly Hondula - would also be valuable.
