### 2019-04-08 #292 @melvidoni
### Editor checks:

---

#### Editor comments

Thanks for your submission @noamross!

As the output from `goodpractice` is perfect I'll start finding reviewers for this.

```
♥ Ho-ho! Slick package! Keep up the stunning work!
```

---

Reviewers: TBD
Due date: TBD
