### 2019-04-15 #292 @pachamaltese
Hi @noamross, @melvidoni and @xavi-rp
I started to install the package. One immediate thing to fix is to add the ROpenSci badge so please paste this code next to CI badge and regenerate the README:
