### 2019-04-15 ecohealthalliance/citesdb#1 @pachamaltese
Connection is garbage-collected, use dbDisconnect() to avoid this
Here's `devtools::check()` output:

```r
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/reviews/citesdb/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (16.9s)
   Warning in eval(`_inherit`, env, NULL) :
     Connection is garbage-collected, use dbDisconnect() to avoid this.
   Error: C stack usage  7971348 is too close to the limit
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```

and here's `sessionInfo()`:

```r
> sessionInfo()
R version 3.5.1 (2018-07-02)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Pop!_OS 18.04 LTS

Matrix products: default
BLAS: /usr/lib/x86_64-linux-gnu/openblas/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/libopenblasp-r0.2.20.so

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                  LC_ADDRESS=C
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] R.utils_2.8.0     R.oo_1.22.0       R.methodsS3_1.7.1 rappdirs_0.3.1
[5] MonetDBLite_0.6.0

 [1] Rcpp_1.0.1        pillar_1.3.1      compiler_3.5.1    prettyunits_1.0.2
 [5] remotes_2.0.2     tools_3.5.1       testthat_2.0.1    digest_0.6.18
 [9] packrat_0.5.0     pkgbuild_1.0.2    pkgload_1.0.2     memoise_1.1.0
[13] tibble_2.1.1      pkgconfig_2.0.2   rlang_0.3.3       DBI_1.0.0
[17] cli_1.1.0         rstudioapi_0.9.0  curl_3.3          withr_2.1.2
[21] dplyr_0.8.0.1     httr_1.4.0        fs_1.2.6          desc_1.2.0
[25] devtools_2.0.1    rprojroot_1.3-2   tidyselect_0.2.5  glue_1.3.1
[29] R6_2.4.0          processx_3.3.0    sessioninfo_1.1.1 purrr_0.3.2
[33] callr_3.2.0       citesdb_0.1.0     magrittr_1.5      usethis_1.4.0
[37] backports_1.1.3   codetools_0.2-15  ps_1.3.0          assertthat_0.2.1
[41] crayon_1.3.4
```
