### 2019-04-16 #292 @pachamaltese
@eveskew Hi! last night I sent you a PR but I just address the problem, I'll try to find a fix with the DB connection.
