### 2019-04-16 ecohealthalliance/citesdb#1 @noamross
Hi @pachamaltese thanks for this.  Can you clarify the issue?  I _think_ you mean that:

- You get `Connection is garbage-collected, use dbDisconnect()` as a warning while using the package?
- You can install the package, but you're getting that `C stack usage` error when running R CMD check?
