### 2019-04-16 ecohealthalliance/citesdb#1 @pachadotdev
@noamross Yes, the `dbDisconnect()` returns a warning and then `R CMD check` fails because of the C error. Let me try on another machine, or maybe a droplet.
