### 2019-04-16 ecohealthalliance/citesdb#2 @pachamaltese
Fix testhat path
HI @noamross

This is `devtools::test()` result:

```r
==> devtools::test()

Loading citesdb
Testing citesdb
✔ | OK F W S | Context
✔ |  1       | Download [157.3 s]
✔ |  6       | Tables [0.5 s]
✔ |  9       | Connection [3.1 s]
✔ |  1       | Data [68.1 s]
✖ |  0 1     | lints [4.1 s]
────────────────────────────────────────────
test-lint.R:5: failure: Package Style
Not lint free
R/connect.R:149:14: style: Variable and function names should be all lowercase.
    observer$connectionClosed("CITESDB", "citesdb")
             ^~~~~~~~~~~~~~~~
R/connection-pane.R:34:14: style: Variable and function names should be all lowercase.
    observer$connectionOpened(
             ^~~~~~~~~~~~~~~~
R/connection-pane.R:65:64: style: Variable and function names should be all lowercase.
                        paste("SELECT * FROM", table, "LIMIT", rowLimit))
                                                               ^~~~~~~~
R/connection-pane.R:85:14: style: Variable and function names should be all lowercase.
    observer$connectionUpdated("CITESDB", "citesdb", "")
             ^~~~~~~~~~~~~~~~~
R/download.R:117:52: warning: Do not use absolute paths.
    paste0("https://api.github.com/repos/", repo, "/releases")
                                                   ^~~~~~~~~
────────────────────────────────────────────

══ Results ═════════════════════════════════
Duration: 233.2 s

OK:       17
Failed:   1
Warnings: 0
Skipped:  0
```
