### 2019-04-16 ecohealthalliance/citesdb#2 @noamross
Thanks. These I think these are all artifacts, as the function names are part of the RStudio API, and the `download.R` value isn't a file path.  I'll add `#nolint` to these.
