### 2019-04-16 ropensci/citesdb#4 @pachamaltese
goodpractice output shows test coverage should be extended
Please run `goodpractice::gp()` to check how to expand test coverage.

Here's my output:

```r
> goodpractice::gp()
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP citesdb ───────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 0%
    of code lines are covered by test cases.

    R/connect.R:3:NA
    R/connect.R:4:NA
    R/connect.R:5:NA
    R/connect.R:7:NA
    R/connect.R:12:NA
    ... and 208 more lines

─────────────────────────────────────────────────────────────────────────────────────────────────
Warning messages:
1: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for cyclomatic complexity failed.
2: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for rcmdcheck failed.
```
