### 2019-04-16 #292 @pachamaltese
* [x]  **Functionality:** Any functional claims of the software been confirmed. _Checked yes but please check the vignettes examples!_

* [ ]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines _I

I enjoyed reviewing the `citesdb` package. I found it easy to use.

Below I've highlighted some observations that might make this package even better.

**Package Metadata**

* I appreciate that the authors included the ORCID ids!
* What I like from the metadata is that `print(citation("citesdb"), bibtex = TRUE)` works ok and returns a proper BibTeX entry with no warnings.

**Package Documentation**

* The GIF is a nice touch!

**Vignettes**

* The vignettes couldn't be generated. The problem is described here: https://github.com/ecohealthalliance/citesdb/issues/1
* While I get "no vignettes found" due to the errors above, I read the Rmd file instead of the rendered vignette. The bibliography is well formatted and I like that you used the brackets in a correct way. This shall ease proper citation and identifying sources.
* With respect to the vignettes contents I'd change the installation part a bit and use `devtools::install_github()`, and also `install.packages()` if you plan to submit to CRAN later.
* I consider really useful that you warn on disk space usage, and also the vignette covers the case when the local database is missing or removed, and how to fix that problem.
* The examples can be improved a bit. For example, the part that says "Note that running `collect()` on all of `cites_shipments()` will load a >3 GB data frame into memory!" is something that I would prefer to re-state the example than to include that kind of warning. For example, and given that the package can be used with dplyr, I would obtain an additional benefit from the SQL connections and use dplyr functions to return just what I want instead of loading 3GB in RAM and then filtering (i.e. see https://db.rstudio.com/dplyr/)
* In the same line with performant examples, I'd run something like this
  ```r
  library(citesdb)

  con <- cites_db()
  DBI::dbListTables(con)

  cites_shipments_db <- tbl(con, "cites_shipments")

  cites_shipments_db

  example_query <- cites_shipments_db %>%
  select(Year, Appendix, Taxon, Exporter) %>%
  filter(Appendix == "II" & Exporter == "DE")

  example_query

  example_query_tibble <- as_tibble(example_query)

  example_query_tibble
  ```
* The integration with MonetDBLite is well covered and explains the existing limitations.
* If there's something I really like from this vignette is that even when it is full of latin names and technical words, I was able to understand it by being just concentrated. Well done!

**Function Documentation**

* The functions are well covered and I had no problem reading the documentation.

#### Functionality / Implementation

**External Dependencies**

[Section 1.11](https://ropensci.github.io/dev_guide/building.html#package-dependencies) of rOpenSci's packaging guides reinforces the good practice of trying to limit
external dependencies where possible.
This is a good practice not only for the health and longevity of your package but also for marketing purposes; sometimes people working on servers or behind proxies
will have limited ability to install packages so fewer dependencies may enable more users.
I think this aspect is ok and I wouldn't touch the dependencies.

**Readability / Modularity**

* With respect to a personal taste, I'd love to see the `glue` package in use here. It can make the code more readable.
* I sent a PR where I applied styler to the code to make it more neat https://github.com/ecohealthalliance/citesdb/pull/3 but two functions need additional review since styler returns an error. The rest of the functions got some minimal re-arrangements mostly related to spacing and line width.

**Technical Implementation Details**

* I tested on Ubuntu 18.04. I would like to get a Windows laptop and check the package from there to see if I would need R Tools or something special to run the package.
* `check_status` should be renamed since it's too generic and `cites_check_status` or similar should be used.

**Testing**

* There is an error and warning when running `devtools::check()` that I reported at https://github.com/ecohealthalliance/citesdb/issues/1
* With `devtools::test()` One test fails. Please check my report here: https://github.com/ecohealthalliance/citesdb/issues/2
* I did run `goodpractice::gp()`. Please check the results here: https://github.com/ecohealthalliance/citesdb/issues/4. Test coverage should be expanded perhaps with some trivial sanity checks.

**Error Messages**

* The error messages are helpful and have no idea to improve then since the texts are complete and informative

Like I stated above, there are some fixes that are needed and may improve the package a lot.

The code is easy to read, without strange lines, and my overall impression of it is positive.

65% of test coverage seems to be very low. By adding some elementary checks, this number can be increased a lot.

Unlike my 1st submission to rOpenSci, most of the functions in citedb have adecuate names and can be easily distinguished as those names are not generic (i.e. "filter" or "download").
