### 2019-04-23 #292 @noamross
Note to reviewers: The *MonetDBLite* package has been (hopefully temporarily) archived on CRAN, so I have switched the *citesdb* to using a stable GitHub version of *MonetDBLite* via the `Remotes:` field in the DESCRIPTION file.
