### 2019-05-01 ecohealthalliance/citesdb#1 @xavi-rp
I'm having the same error when R CMD check. However, I have no problems while using the package (`cites_disconnect()`)

```
==> devtools::check()

Updating citesdb documentation
Loading citesdb
Writing NAMESPACE
── Building ──────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
───────────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Downloads/citesdb-master/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (44.2s)
   Warning in name :
     Connection is garbage-collected, use dbDisconnect() to avoid this.
   Error: C stack usage  7969408 is too close to the limit
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```

```
> sessionInfo()
R version 3.5.3 RC (2019-03-04 r76210)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.4

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.5/Resources/lib/libRlapack.dylib

locale:
[1] ca_ES.UTF-8/ca_ES.UTF-8/ca_ES.UTF-8/C/ca_ES.UTF-8/ca_ES.UTF-8

attached base packages:

other attached packages:
[1] shiny_1.3.2    reprex_0.2.1   citesdb_0.1.0  usethis_1.4.0  devtools_2.0.1

 [1] tidyselect_0.2.5  xfun_0.6          remotes_2.0.2     purrr_0.3.2
 [5] testthat_2.1.1    miniUI_0.1.1.1    htmltools_0.3.6   yaml_2.2.0
 [9] rlang_0.3.4       pkgbuild_1.0.3    later_0.8.0       R.oo_1.22.0
[13] MonetDBLite_0.6.0 pillar_1.3.1      glue_1.3.1        withr_2.1.2
[17] DBI_1.0.0         R.utils_2.8.0     rappdirs_0.3.1    dbplyr_1.4.0
[21] sessioninfo_1.1.1 stringr_1.4.0     commonmark_1.7    R.methodsS3_1.7.1
[25] codetools_0.2-16  memoise_1.1.0     evaluate_0.13     knitr_1.22
[29] callr_3.2.0       httpuv_1.5.1      ps_1.3.0          Rcpp_1.0.1
[33] xtable_1.8-4      clipr_0.5.0       promises_1.0.1    backports_1.1.4
[37] desc_1.2.0        pkgload_1.0.2     jsonlite_1.6      mime_0.6
[41] fs_1.2.6          digest_0.6.18     stringi_1.4.3     processx_3.3.0
[45] dplyr_0.8.0.1     rprojroot_1.3-2   cli_1.1.0         tools_3.5.3
[49] magrittr_1.5      tibble_2.1.1      crayon_1.3.4      whisker_0.3-2
[53] pkgconfig_2.0.2   xml2_1.2.0        prettyunits_1.0.2 assertthat_0.2.1
[57] rmarkdown_1.12    roxygen2_6.1.1    httr_1.4.0        rstudioapi_0.10
[61] R6_2.4.0          compiler_3.5.3
```
