### 2019-05-04 ecohealthalliance/citesdb#1 @xavi-rp
Hi,
Apparently the problem when `R CMD check` is that there is something recursive when building vignettes that makes it crash with the `C stack usage` error .

These are the results of `devtools::check(args = c('--no-build-vignettes'), build_args = c('--no-build-vignettes','--no-build-vignettes'))`

```
==> devtools::check(args = c('--no-build-vignettes'), build_args = c('--no-build-vignettes','--no-build-vignettes'))

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Documents/citesdb/DESCRIPTION’ ...
─  preparing ‘citesdb’:
─  building ‘citesdb_0.1.0.tar.gz’

── Checking ────────────────────────── citesdb ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_REMOTE_: FALSE
● _R_CHECK_CRAN_INCOMING_       : FALSE
● _R_CHECK_FORCE_SUGGESTS_      : FALSE
── R CMD check ─────────────────────────────────────────────────────────────────
─  using log directory ‘/Users/xavi_rp/Documents/citesdb.Rcheck’
─  using R version 3.5.3 RC (2019-03-04 r76210)
─  using platform: x86_64-apple-darwin15.6.0 (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --no-build-vignettes --as-cran’
✔  checking for file ‘citesdb/DESCRIPTION’
─  this is package ‘citesdb’ version ‘0.1.0’
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (2.8s)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names ...
✔  checking for sufficient/correct file permissions
✔  checking serialization versions
✔  checking whether package ‘citesdb’ can be installed (1.6s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking top-level files
✔  checking for left-over files
✔  checking index information
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (584ms)
✔  checking whether the package can be loaded with stated dependencies (495ms)
✔  checking whether the package can be unloaded cleanly (497ms)
✔  checking whether the namespace can be loaded with stated dependencies (488ms)
✔  checking whether the namespace can be unloaded cleanly (546ms)
✔  checking dependencies in R code (537ms)
✔  checking S3 generic/method consistency (1.2s)
✔  checking replacement functions (523ms)
✔  checking foreign function calls (523ms)
✔  checking R code for possible problems (2.8s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (515ms)
✔  checking for code/documentation mismatches (1.6s)
✔  checking Rd \usage sections (1.3s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
W  checking files in ‘vignettes’ ...
   Files in the 'vignettes' directory but no files in 'inst/doc':
     ‘citations.bib’, ‘exploring-cites-trade-data-with-citesdb.Rmd’,
       ‘figures/cites_pane.gif’, ‘figures/eha-footer.png’
✔  checking examples (1.8s)
   ** found \donttest examples: check also with --run-donttest
✔  checking for unstated dependencies in vignettes ...
W  checking package vignettes in ‘inst/doc’
   Package vignette without corresponding PDF/HTML:
      ‘exploring-cites-trade-data-with-citesdb.Rmd’

─  checking running R code from vignettes ...
      ‘exploring-cites-trade-data-with-citesdb.Rmd’ using ‘UTF-8’ ... OK
    NONE
─  checking re-building of vignette outputs ... SKIPPED

   See
     ‘/Users/xavi_rp/Documents/citesdb.Rcheck/00check.log’
   for details.

── R CMD check results ────────────────────────────────────── citesdb 0.1.0 ────
Duration: 19.7s

❯ checking files in ‘vignettes’ ... WARNING
  Files in the 'vignettes' directory but no files in 'inst/doc':
    ‘citations.bib’, ‘exploring-cites-trade-data-with-citesdb.Rmd’,
      ‘figures/cites_pane.gif’, ‘figures/eha-footer.png’

❯ checking package vignettes in ‘inst/doc’ ... WARNING
  Package vignette without corresponding PDF/HTML:
     ‘exploring-cites-trade-data-with-citesdb.Rmd’

0 errors ✔ | 2 warnings ✖ | 0 notes ✔
Error: R CMD check found WARNINGs
Execution halted

Exited with status 1.
```

I think that the problem might be related to the use of `MonetDBLite::monetdblite_shutdown()` in ` cites_disconnect` because, if there is no connection, it produces an error. Using `DBI::dbDisconnect()` instead, it might solve the problem (but not sure...).
