### 2019-05-04 ecohealthalliance/citesdb#1 @noamross
Thanks for this. Yes, I also suspect it has something to do with the shutdown, especially the `reg.finalizer` call at https://github.com/ecohealthalliance/citesdb/blob/master/R/connect.R#L154 which disconnects on abort.  I hope to tackle all the review issues in one sprint - will this prevent you from completing the review?
