### 2019-05-05 #292 @xavi-rp
Thanks for giving me the opportunity to review *citesdb*. I think it's a very nice and useful package. I've enjoyed (and learned!) a lot reviewing it.
Please take my comments only as suggestions as I think it's already very well written.

    - *It gives an error while running `devtools::check()` (see comments below). However, reading the Rmd file and running the examples in it manually, it can be seen that they work perfectly.*

Estimated hours spent reviewing: 8 (a lot of time spent to try to understand the problem with the vignettes)

---

I find this package easy to install/use and very useful to explore and use the data from the CITES Trade DB, so congrats to the authors!

It's very useful the inclusion of the metadata functions. And also the gif!!!

The main problem that I found is in the vignette building while running `devtools::check()`. See https://github.com/ecohealthalliance/citesdb/issues/1#issuecomment-489351239

In `devtools::test()`, the lints test fails. Already reported by @pachamaltese here: https://github.com/ecohealthalliance/citesdb/issues/2
These are the results of checking the lints with `lintr::lint_package()` (most of them are probably irrelevant; except maybe #13)

```
> library(lintr)
> as.data.frame(lint_package())

              filename line_number column_number    type
1          R/connect.R         114            13   style
2          R/connect.R         121            13   style
3          R/connect.R         128            13   style
4          R/connect.R         149            14   style
5  R/connection-pane.R          34            14   style
6  R/connection-pane.R          65            64   style
7  R/connection-pane.R          85            14   style
8         R/download.R          41             7   style
9         R/download.R          42             5   style
10        R/download.R          46             3   style
11        R/download.R          49             5   style
12        R/download.R          59             3   style
13        R/download.R         117            52 warning
14        R/onattach.R          24            13   style
15        R/onattach.R          25             5   style
16        R/onattach.R          42             7   style
17        R/onattach.R          43             5   style
18        R/onattach.R          70            22   style
                                                message
1  Variable and function names should be all lowercase.
2  Variable and function names should be all lowercase.
3  Variable and function names should be all lowercase.
4  Variable and function names should be all lowercase.
5  Variable and function names should be all lowercase.
6  Variable and function names should be all lowercase.
7  Variable and function names should be all lowercase.
8  Variable and function names should be all lowercase.
9  Variable and function names should be all lowercase.
10 Variable and function names should be all lowercase.
11 Variable and function names should be all lowercase.
12 Variable and function names should be all lowercase.
13                           Do not use absolute paths.
14 Variable and function names should be all lowercase.
15 Variable and function names should be all lowercase.
16 Variable and function names should be all lowercase.
17 Variable and function names should be all lowercase.
18 Variable and function names should be all lowercase.
                                                                           line
1                          as_tibble(dbReadTable(cites_db(), "cites_metadata"))
2                             as_tibble(dbReadTable(cites_db(), "cites_codes"))
3                           as_tibble(dbReadTable(cites_db(), "cites_parties"))
4                               observer$connectionClosed("CITESDB", "citesdb")
5                                                    observer$connectionOpened(
6                             paste("SELECT * FROM", table, "LIMIT", rowLimit))
7                          observer$connectionUpdated("CITESDB", "citesdb", "")
8                                     if (dbExistsTable(cites_db(), tblname)) {
9                                            dbRemoveTable(cites_db(), tblname)
10               dbCreateTable(cites_db(), tblname, fields = cites_field_types)
11                                                                   dbExecute(
12   dbWriteTable(cites_db(), "cites_status", make_status_table(version = ver),
13                   paste0("https://api.github.com/repos/", repo, "/releases")
14                                        for (t in dbListTables(cites_db())) {
15                                                 dbRemoveTable(cites_db(), t)
16                          if (dbExistsTable(cites_db(), "cites_shipments") &&
17                                 dbExistsTable(cites_db(), "cites_status")) {
18                       suppressMessages(dbWriteTable(cites_db(), tblnames[i],
                  linter
1      camel_case_linter
2      camel_case_linter
3      camel_case_linter
4      camel_case_linter
5      camel_case_linter
6      camel_case_linter
7      camel_case_linter
8      camel_case_linter
9      camel_case_linter
10     camel_case_linter
11     camel_case_linter
12     camel_case_linter
13 absolute_paths_linter
14     camel_case_linter
15     camel_case_linter
16     camel_case_linter
17     camel_case_linter
18     camel_case_linter
```

The spelling errors found after running `spelling::spell_check_package()` and `spelling::spell_check_files("README.Rmd")` are irrelevant.

#### Other minor comments:
- I would like to know more easily where the DB is stored; for instance when calling `cites_status()`

- If you do `cites_db_delete()` and then again `cites_db_download()`, the Size on Disk becomes 2.3 Gb instead of 1.1Gb. I don't know why this happens, but it's not nice...

- `cites_pane()` `@examples` L29-30. It should be in the same line (or using {}).

- I would prefer single examples for `cites_metadata()`, `cites_codes()` and `cites_parties()`

- Use Title Case in the package title

Congrats again to the authors!!
