### 2019-05-06 #292 @noamross
Thank you for your careful reviews, @pachamaltese and @xavi-rp! We'll get to work on these.
