### 2019-05-08 ecohealthalliance/citesdb#1 @noamross
Hi all, I haven't been able to reproduce the recursion error in building the vignette, but I've made some changes that affect how we disconnect from the database the _might_ fix it on your end.  Would you be able to try to check or run the vignette on the current `ro-review` branch and let me know if you see different behavior?
