### 2019-05-09 ecohealthalliance/citesdb#1 @pachadotdev
@noamross sorry for the late reply, I have this problem :(

```
> devtools::check()
Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Warning in .local(conn, ...) : already disconnected
Warning in monetdb_embedded_disconnect(conn@connenv$conn) :
  Connection was already disconnected.
Local CITES database empty or corrupt. Download with cites_db_download()
Writing NAMESPACE
── Building ────────────────────────────────────────────────────────────────────────────────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/citesdb/DESCRIPTION’
─  preparing ‘citesdb’:
E  creating vignettes (2.4s)
   Quitting from lines 56-57 (exploring-cites-trade-data-with-citesdb.Rmd)
   Error: processing vignette 'exploring-cites-trade-data-with-citesdb.Rmd' failed with diagnostics:
   Local citesdb database is locked by another R session.
    Try closing or running cites_disconect() in that session.
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
```

but the line with the problem runs perfect from console (`if (!citesdb::cites_status()) citesdb::cites_db_download()`)
