### 2019-05-09 ecohealthalliance/citesdb#1 @xavi-rp
Hi,
Keep having an error...

```
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ─────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
──────────────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Documents/citesdb_ro_review/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (43.8s)
   Warning in if (is.null(class2Def)) FALSE else if (!.identC(class(class2Def),  :
     Connection is garbage-collected, use dbDisconnect() to avoid this.
   Error: C stack usage  7969392 is too close to the limit
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```

But no errors when checking with ` --no-build-vignettes `

```
==> devtools::check(args = c('--no-build-vignettes'), build_args = c('--no-build-vignettes','--no-build-vignettes'))

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ──────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
───────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Documents/citesdb_ro_review/DESCRIPTION’ ...
─  preparing ‘citesdb’:
─  building ‘citesdb_0.1.0.tar.gz’

── Checking ──────────── citesdb ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_REMOTE_: FALSE
● _R_CHECK_CRAN_INCOMING_       : FALSE
● _R_CHECK_FORCE_SUGGESTS_      : FALSE
── R CMD check ─────────────────────────────────────────────────────────────────
─  using log directory ‘/Users/xavi_rp/Documents/citesdb.Rcheck’
─  using R version 3.5.3 RC (2019-03-04 r76210)
─  using platform: x86_64-apple-darwin15.6.0 (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --no-build-vignettes --as-cran’
✔  checking for file ‘citesdb/DESCRIPTION’
─  this is package ‘citesdb’ version ‘0.1.0’
─  package encoding: UTF-8
✔  checking package namespace information ...
✔  checking package dependencies (2.3s)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names ...
✔  checking for sufficient/correct file permissions
✔  checking serialization versions ...
✔  checking whether package ‘citesdb’ can be installed (1.6s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking top-level files
✔  checking for left-over files
✔  checking index information
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (604ms)
✔  checking whether the package can be loaded with stated dependencies (499ms)
✔  checking whether the package can be unloaded cleanly (572ms)
✔  checking whether the namespace can be loaded with stated dependencies (527ms)
✔  checking whether the namespace can be unloaded cleanly (566ms)
✔  checking dependencies in R code (542ms)
✔  checking S3 generic/method consistency (1.2s)
✔  checking replacement functions (520ms)
✔  checking foreign function calls (548ms)
✔  checking R code for possible problems (2.8s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (498ms)
✔  checking for code/documentation mismatches (1.5s)
✔  checking Rd \usage sections (1.2s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
W  checking files in ‘vignettes’ ...
   Files in the 'vignettes' directory but no files in 'inst/doc':
     ‘citations.bib’, ‘exploring-cites-trade-data-with-citesdb.Rmd’,
       ‘figures/cites_pane.gif’, ‘figures/eha-footer.png’
✔  checking examples (3.7s)
   ** found \donttest examples: check also with --run-donttest
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✔  Comparing ‘spelling.Rout’ to ‘spelling.Rout.save’ ... OK (568ms)
✔  Running ‘testthat.R’ [116s/275s]
✔  checking for unstated dependencies in vignettes ...
W  checking package vignettes in ‘inst/doc’
   Package vignette without corresponding PDF/HTML:
      ‘exploring-cites-trade-data-with-citesdb.Rmd’

─  checking running R code from vignettes ...
      ‘exploring-cites-trade-data-with-citesdb.Rmd’ using ‘UTF-8’ ... OK
    NONE
─  checking re-building of vignette outputs ... SKIPPED

   See
     ‘/Users/xavi_rp/Documents/citesdb.Rcheck/00check.log’
   for details.

── R CMD check results ────────────────────────────────────── citesdb 0.1.0 ────
Duration: 4m 57.6s

❯ checking files in ‘vignettes’ ... WARNING
  Files in the 'vignettes' directory but no files in 'inst/doc':
    ‘citations.bib’, ‘exploring-cites-trade-data-with-citesdb.Rmd’,
      ‘figures/cites_pane.gif’, ‘figures/eha-footer.png’

❯ checking package vignettes in ‘inst/doc’ ... WARNING
  Package vignette without corresponding PDF/HTML:
     ‘exploring-cites-trade-data-with-citesdb.Rmd’

0 errors ✔ | 2 warnings ✖ | 0 notes ✔
Error: R CMD check found WARNINGs
Execution halted

```

I'm sorry... I can't find where my error comes from...

This is my session info:
```
> sessionInfo()
R version 3.5.3 RC (2019-03-04 r76210)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.4

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.5/Resources/lib/libRlapack.dylib

locale:
[1] ca_ES.UTF-8/ca_ES.UTF-8/ca_ES.UTF-8/C/ca_ES.UTF-8/ca_ES.UTF-8

attached base packages:
[1] stats     graphics  grDevices
[4] utils     datasets  methods
[7] base

[1] compiler_3.5.3 tools_3.5.3
[3] yaml_2.2.0
```
