### 2019-05-09 ecohealthalliance/citesdb#1 @noamross
Hmm, still hunting this down and unable to recreate this, even though I've extended my CI to include older R versions.  I've made one small change just pushed to the `ro-review` branch, running `gc()` before and after database connection, which again, _may_ help this.  Could try testing again with this?

@pachamaltese, I think your error comes from having a connection open when you run check? (Often having the package loaded in the console when running in the check pane).  Unfortunately a limit of MonetDBLite (to be fixed in its successor package!) is that only one session can be opened at once.  This is a limitation that I'll document more explicitly.
