### 2019-05-09 ecohealthalliance/citesdb#1 @pachadotdev
@noamross I tested after rebooting, let me try on a fresh droplet
