### 2019-05-09 ecohealthalliance/citesdb#1 @xavi-rp
Hi,
Still there...
The warning just before crashing is different, but the result is the same.
Tomorrow I'll check it in another computer, let's see if something different happens there...

```
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Documents/ro_review/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (42.6s)
   Warning in self$draw_panel :
     Connection is garbage-collected, use dbDisconnect() to avoid this.
   Error: C stack usage  7969312 is too close to the limit
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```
