### 2019-05-09 ecohealthalliance/citesdb#1 @noamross
I don't want to impose, but would either of you see be able to if this occurs with R 3.6.0 in your environments?
