### 2019-05-09 ecohealthalliance/citesdb#1 @pachadotdev
@noamross I'm testing it on a fresh droplet (I created a new droplet with the rstudio image from marketplace), thinking that the vanilla ubuntu droplet could be a problem, it returns this now

```
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ───────────────────────────────────────────────────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/rstudio/citesdb/DESCRIPTION’
─  preparing ‘citesdb’:
✔  checking DESCRIPTION meta-information ... OK
E  creating vignettes (8.4s)
   Warning in if (any(is.na(x))) stop("assert_that: missing values present in assertion",  :
     Connection is garbage-collected, use dbDisconnect() to avoid this.
   Quitting from lines 167-170 (exploring-cites-trade-data-with-citesdb.Rmd)
   Error: processing vignette 'exploring-cites-trade-data-with-citesdb.Rmd' failed with diagnostics:
   there is no package called 'rcites'
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```
