### 2019-05-11 ecohealthalliance/citesdb#1 @xavi-rp
Hi,
Checked with R3.6.0 and the error persists. It looks like `cites_disconect()`is not able to close the open connection.

```
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Registered S3 method overwritten by 'R.oo':
  method        from
  throw.default R.methodsS3
Writing NAMESPACE
── Building ────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────
✔  checking for file ‘/Users/xavi_rp/Documents/ro_review/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (3.2s)
   --- re-building ‘exploring-cites-trade-data-with-citesdb.Rmd’ using rmarkdown
   Registered S3 method overwritten by 'R.oo':
     method        from
     throw.default R.methodsS3
   Quitting from lines 56-57 (exploring-cites-trade-data-with-citesdb.Rmd)
   Error: processing vignette 'exploring-cites-trade-data-with-citesdb.Rmd' failed with diagnostics:
   Local citesdb database is locked by another R session.
    Try closing or running cites_disconect() in that session.
   --- failed re-building ‘exploring-cites-trade-data-with-citesdb.Rmd’

   SUMMARY: processing the following file failed:
     ‘exploring-cites-trade-data-with-citesdb.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```

Although it says 'Local citesdb database is locked by another R session', I have no other open R session.
