### 2019-05-19 ecohealthalliance/citesdb#1 @noamross
I just pushed up another change, that I hope you can test, though I am admittedly kind of mashing buttons as I still can't reproduce the C Stack error.
