### 2019-05-19 ecohealthalliance/citesdb#1 @xavi-rp
Hi,
Just tried and the error is still there. Just to check different (and maybe stupid) things, I put `  DBI::dbDisconnect(db, shutdown = TRUE)` outside the if statement, like this:
```
cites_disconnect_ <- function(environment = cites_cache) { # nolint
  db <- mget("cites_db", envir = cites_cache, ifnotfound = NA)[[1]]
  DBI::dbDisconnect(db, shutdown = TRUE)
  if (inherits(db, "DBIConnection")) {
    gc(verbose = FALSE)
    DBI::dbDisconnect(db, shutdown = TRUE)
  }
  observer <- getOption("connectionObserver")
  if (!is.null(observer)) {
    observer$connectionClosed("CITESDB", "citesdb")
  }
}
```

The error message is now different: 'unable to find an inherited method for function 'dbDisconnect' for signature '"logical"''. And nothing about the C error...
It seems like `db <- mget("cites_db", envir = cites_cache, ifnotfound = NA)[[1]]` is not a connection and, therefore, the open connection can't be closed.
Not sure, but this might help (hopefully)...
