### 2019-05-19 ecohealthalliance/citesdb#1 @pachadotdev
@xavi-rp @noamross yes, same here:

```
==> devtools::check()

Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
Writing NAMESPACE
── Building ────────────────────────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/citesdb/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (13.3s)
   Warning: Connection is garbage-collected, use dbDisconnect() to avoid this.
   Error: C stack usage  7971268 is too close to the limit
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... eval -> with_envvar -> force -> do.call -> <Anonymous>
Execution halted

Exited with status 1.
```
