### 2019-05-20 ecohealthalliance/citesdb#1 @noamross
And do you both get the C stack error knitting the vignette outside R CMD check? (In a fresh session with no other sessions open)
