### 2019-05-20 ecohealthalliance/citesdb#1 @pachadotdev
Let me check again. In the previous commit that didn't happen outside
knit/test.
