### 2019-05-20 ecohealthalliance/citesdb#1 @noamross
In either case, can you both add the following line to the first block in the vignette and attempt to build? This should dump the environment to a file (`citesdb_dump.rda`) in your home directory (you can choose a different path, but it needs to be an absolute path as the directory where the package is built is temporary).  If you then send me the file I should be able to inspect it.

```
options(error = function() dump.frames(dumpto = "~/citesdb_dump", to.file = TRUE))
```
