### 2019-05-21 ecohealthalliance/citesdb#1 @xavi-rp
To your first question, I get the C stack error also when knitting the vignette outside R CMD check.
Second, attached the 'cites_dump.rda' file.

Hope you find the problem!!

[citesdb_dump.rda.zip](https://github.com/ecohealthalliance/citesdb/files/3201201/citesdb_dump.rda.zip)
