### 2019-05-21 ecohealthalliance/citesdb#1 @noamross
Eureka! I believe I have found it.  It turns out that that the recursion error resulting in the C Stack error was due to not properly handling the **rcites** token.  I've pushed changes, again to the `ro-review` branch. If this works for you, we'll merge, submit the revision and our response to reviewers.
