### 2019-05-22 ecohealthalliance/citesdb#1 @pachadotdev
I have a problem:

```r
> devtools::check()
Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
CITES database status:
Time Imported: 2019-05-09 00:55:05
Version: 0.1.0
Number of Records: 20,215,030
Size on Disk: 1.1 Gb
Location on Disk: /home/pacha/.local/share/citesdb
Writing NAMESPACE
── Building ──────────────────────────────────────────────────────────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
───────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/citesdb/DESCRIPTION’ ...
─  preparing ‘citesdb’:
E  creating vignettes (2.6s)
   Quitting from lines 56-57 (exploring-cites-trade-data-with-citesdb.Rmd)
   Error: processing vignette 'exploring-cites-trade-data-with-citesdb.Rmd' failed with diagnostics:
   Local citesdb database is locked by another R session.
    Try closing or running cites_disconect() in that session.
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
```
