### 2019-05-22 ecohealthalliance/citesdb#1 @pachadotdev
also

```r
> library(citesdb)
> cites_disconect()
Error in cites_disconect() : could not find function "cites_disconect"
```
