### 2019-05-22 ecohealthalliance/citesdb#1 @eveskew
Hi @pachamaltese. Thanks for the continued help. With regards to your most recent query, the function is actually `cites_disconnect()` rather than `cites_disconect()`. I'll fix that. And when you're attempting to build the package, have you tried starting a new R session and **not** loading `citesdb`? If you've been using any of the package functionality prior to trying to build then it's expected that you'll get a locked database.
