### 2019-05-22 ecohealthalliance/citesdb#1 @pachadotdev
@eveskew I saw the last push, it fixed the `Try closing or running cites_disconect() in that session.` message but it returns this warning:
```
> cites_disconnect()
Warning message:
In .Internal(gc(verbose, reset, full)) :
  Connection is garbage-collected, use dbDisconnect() to avoid this.
```
