### 2019-05-22 ecohealthalliance/citesdb#1 @noamross
Ah, the `gc()` call was something that I included when I was trying to chase the recursion error.  I just removed it in the last push.  Do you get that same warning?
