### 2019-05-22 ecohealthalliance/citesdb#1 @noamross
Any chance you can you get on the RO slack? I need to squash this bug once and for all.
