### 2019-05-22 ecohealthalliance/citesdb#1 @pachadotdev
@noamross `e551139` works

### Tests

```r
> library(citesdb)
CITES database status:
Time Imported: 2019-05-22 18:19:11
Version: 0.1.0
Number of Records: 20,215,030
Size on Disk: 1.1 Gb
Location on Disk: /home/pacha/.local/share/citesdb
> devtools::check()
Updating citesdb documentation
Writing NAMESPACE
Loading citesdb
CITES database status:
Time Imported: 2019-05-22 18:19:11
Version: 0.1.0
Number of Records: 20,215,030
Size on Disk: 1.1 Gb
Location on Disk: /home/pacha/.local/share/citesdb
Writing NAMESPACE
── Building ──────────────────────────────────────────────────────────────────────────────── citesdb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
───────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/citesdb/DESCRIPTION’ ...
─  preparing ‘citesdb’:
✔  creating vignettes (8.4s)
─  building ‘citesdb_0.1.0.tar.gz’

── Checking ──────────────────────────────────────────────────────────────────────────────── citesdb ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
● _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
● _R_CHECK_CRAN_INCOMING_           : FALSE
● _R_CHECK_FORCE_SUGGESTS_          : FALSE
── R CMD check ────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/tmp/RtmpsGetBX/citesdb.Rcheck’ (336ms)
─  using R version 3.5.3 (2019-03-11)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘citesdb/DESCRIPTION’
─  this is package ‘citesdb’ version ‘0.1.0’
─  package encoding: UTF-8
✔  checking package namespace information ...
✔  checking package dependencies (1.9s)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names ...
✔  checking for sufficient/correct file permissions
✔  checking serialization versions
✔  checking whether package ‘citesdb’ can be installed (2.1s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking ‘build’ directory
✔  checking DESCRIPTION meta-information (374ms)
✔  checking top-level files
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories (458ms)
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (769ms)
✔  checking whether the package can be loaded with stated dependencies (688ms)
✔  checking whether the package can be unloaded cleanly (661ms)
✔  checking whether the namespace can be loaded with stated dependencies (758ms)
✔  checking whether the namespace can be unloaded cleanly (949ms)
✔  checking loading without being on the library search path (876ms)
✔  checking dependencies in R code (803ms)
✔  checking S3 generic/method consistency (1.4s)
✔  checking replacement functions (717ms)
✔  checking foreign function calls (671ms)
✔  checking R code for possible problems (4.3s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (696ms)
✔  checking for code/documentation mismatches (2s)
✔  checking Rd \usage sections (1.6s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
✔  checking examples (4.1s)
   ** found \donttest examples: check also with --run-donttest
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✔  Running ‘spelling.R’
✔  Running ‘testthat.R’ (1.3s)
✔  checking for unstated dependencies in vignettes (1.7s)
✔  checking package vignettes in ‘inst/doc’ ...
✔  checking re-building of vignette outputs (6.1s)

── R CMD check results ───────────────────────────────────────────────────────────── citesdb 0.1.0 ────
Duration: 37.2s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```

### sessionInfo()

```r
> sessionInfo()
R version 3.5.3 (2019-03-11)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Pop!_OS 18.04 LTS

Matrix products: default
BLAS: /usr/lib/x86_64-linux-gnu/openblas/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/libopenblasp-r0.2.20.so

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                  LC_ADDRESS=C
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] citesdb_0.1.0

 [1] Rcpp_1.0.1        dbplyr_1.4.0      pillar_1.4.0      compiler_3.5.3    R.methodsS3_1.7.1
 [6] prettyunits_1.0.2 R.utils_2.8.0     remotes_2.0.4     tools_3.5.3       testthat_2.1.1
[11] digest_0.6.19     pkgbuild_1.0.3    pkgload_1.0.2     MonetDBLite_0.6.0 memoise_1.1.0
[16] tibble_2.1.1      pkgconfig_2.0.2   rlang_0.3.4       DBI_1.0.0         cli_1.1.0
[21] rstudioapi_0.10   commonmark_1.7    xopen_1.0.0       xml2_1.2.0        stringr_1.4.0
[26] roxygen2_6.1.1    withr_2.1.2       dplyr_0.8.1       httr_1.4.0        fs_1.3.1
[31] desc_1.2.0        rappdirs_0.3.1    devtools_2.0.2    rprojroot_1.3-2   tidyselect_0.2.5
[36] glue_1.3.1        R6_2.4.0          processx_3.3.1    rcmdcheck_1.3.3   sessioninfo_1.1.1
[41] purrr_0.3.2       callr_3.2.0       magrittr_1.5      backports_1.1.4   codetools_0.2-16
[46] ps_1.3.0          usethis_1.5.0     assertthat_0.2.1  stringi_1.4.3     crayon_1.3.4
[51] R.oo_1.22.0
```
