### 2019-05-22 ecohealthalliance/citesdb#1 @noamross
After a long and merry (ha!) chase I think we've got this one. Thanks so much and look out for a response to reviewers in the review thread.
