### 2019-05-26 #292 @pachamaltese
@melvidoni I totally agree with this:

> Hi,
> I think that the authors have adressed all my review comments and am happy with their changes. Therefore, I recomend `citesdb` for approving.
> Again, thank you so much to the editor for giving me the opportunity to review such a nice package and congratulations to the authors!!
