### 2019-05-26 #292 @pachamaltese
@melvidoni IMO there's an additional pending: to apply rOpenSci themes to their vignettes
