### 2019-05-29 #292 @noamross
Thanks all very much, I believe the RO theme will be applied automatically upon repository transfer and building on `docs.ropensci.org`.

@pachamaltese and @xavi-rp, we'd like to acknowledge you as `rev` authors in our DESCRIPTION if that's all right.  I believe your ORCiDs are https://orcid.org/0000-0003-2046-0621 (Xavier) and 0000-0003-1017-7574 (Pachá), correct?
