### 2019-05-29 #292 @pachamaltese
Correct id!

You have to download the ropensci theme and edit _pkgdown.yml to create an
ropensci-themed pkgdown site

Check github.com/ropensci/tradestatistics for the details
