### 2019-05-29 #292 @noamross
I've checked off everything except the footer, which I'm leaving off in anticipation of a change in workflow: https://github.com/ropensci/software-review-meta/issues/79, and will be in touch with Stefanie about a blog post.  Thanks all!
