### 2019-05-29 #292 @pachamaltese
@noamross I saw ropensci.github.io/citesdb
It just needs the rOpenSci theme and that's it! Just 2 minutes away from
the best result
