### 2019-05-29 #292 @maelle
@pachamaltese https://docs.ropensci.org/tradestatistics/ and https://dev.ropensci.org/job/tradestatistics/ 😉
