### 2019-05-29 #292 @pachamaltese
Thanks @maelle!! It was different when I submitted

On Wed, May 29, 2019, 2:15 PM Maëlle Salmon <notifications@github.com>
wrote:

> @pachamaltese <https://github.com/pachamaltese> we are about to change
> the guidance actually because @jeroen <https://github.com/jeroen> has
> been working on centrally building all docs websites so authors do not need
> to worry about this. The pkgdown config is used but not for styling since
> rotemplate is used for all. See e.g. docs.ropensci.org/magick.
>
